feat: abstract decision provider
