+++ src/turns/providers/llmDecisionProvider.js
@@ -2,67 +2,51 @@
  * @module LLMDecisionProvider
  */
 
-import { ITurnDecisionProvider } from '../interfaces/ITurnDecisionProvider.js';
-import { assertValidActionIndex } from '../../utils/validationUtils.js';
+import { AbstractDecisionProvider } from './abstractDecisionProvider.js';
 
 /** @typedef {import('../../interfaces/ISafeEventDispatcher.js').ISafeEventDispatcher} ISafeEventDispatcher */
 
 /**
  * @class LLMDecisionProvider
- * @augments ITurnDecisionProvider
+ * @augments AbstractDecisionProvider
  * @description
  *  Decision provider that delegates to an LLM chooser implementation.
  */
-export class LLMDecisionProvider extends ITurnDecisionProvider {
+export class LLMDecisionProvider extends AbstractDecisionProvider {
   /**
+   * Creates a new LLMDecisionProvider.
+   *
    * @param {{
    *  llmChooser: import('../../turns/ports/ILLMChooser').ILLMChooser,
-   *  logger: import('../../interfaces/coreServices').ILogger
+   *  logger: import('../../interfaces/coreServices').ILogger,
    *  safeEventDispatcher: ISafeEventDispatcher
-   * }} deps
+   * }} deps - Constructor dependencies
    */
   constructor({ llmChooser, logger, safeEventDispatcher }) {
-    super();
+    super({ logger, safeEventDispatcher });
     /** @protected @type {import('../../turns/ports/ILLMChooser').ILLMChooser} */
     this.llmChooser = llmChooser;
-    /** @protected @type {import('../../interfaces/coreServices').ILogger} */
-    this.logger = logger;
-    /** @protected @type {ISafeEventDispatcher} */
-    this.safeEventDispatcher = safeEventDispatcher;
   }
 
   /**
-   * @override
+   * Delegates decision making to an LLM chooser.
+   *
    * @async
-   * @param {import('../../entities/entity.js').default} actor
-   * @param {import('../interfaces/ITurnContext.js').ITurnContext} context
-   * @param {import('../dtos/actionComposite.js').ActionComposite[]} actions
-   * @param {AbortSignal} [abortSignal]
-   * @returns {Promise<import('../../turns/interfaces/ITurnDecisionProvider').ITurnDecisionResult>}
+   * @protected
+   * @override
+   * @param {import('../../entities/entity.js').default} actor - Acting entity
+   * @param {import('../interfaces/ITurnContext.js').ITurnContext} context - Turn context
+   * @param {import('../dtos/actionComposite.js').ActionComposite[]} actions - Indexed action list
+   * @param {AbortSignal} [abortSignal] - Optional cancellation signal
+   * @returns {Promise<{ index: number, speech?: string|null, thoughts?: string|null, notes?: string[]|null }>} Result selected by the LLM
    */
-  async decide(actor, context, actions, abortSignal) {
+  async choose(actor, context, actions, abortSignal) {
     const { index, speech, thoughts, notes } = await this.llmChooser.choose({
       actor,
       context,
       actions,
       abortSignal,
     });
-
-    assertValidActionIndex(
-      index,
-      actions.length,
-      'LLMDecisionProvider',
-      actor.id,
-      this.safeEventDispatcher,
-      this.logger,
-      { llmResult: { index, speech, thoughts, notes } }
-    );
-
-    return {
-      chosenIndex: index,
-      speech,
-      thoughts,
-      notes,
-    };
+    return { index, speech, thoughts, notes };
   }
 }
